requirements.txt
python-telegram-bot==20.7
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
requests==2.31.0
torch==2.1.2
torchvision==0.16.2
rembg==2.0.53
diffusers==0.26.3
transformers==4.37.2
accelerate==0.26.1
safetensors==0.4.2
aiohttp==3.9.1
python-dotenv==1.0.0
segment-anything==1.0